chore(refactor): extract simple graphql wrappers (github)
